fix integer overflow issue using size_t

+++ huffman.h
@@ -1,2 +1,2 @@
-unsigned char* encode(char* charset, unsigned int* encoded_data_size);
-char* decode(unsigned char* encoded_data, unsigned int encoded_data_size);
+unsigned char* encode(char* charset, size_t* encoded_data_size);
+char* decode(unsigned char* encoded_data, size_t encoded_data_size);
